+++ simplivity/resources/omnistack_clusters.py
@@ -112,7 +112,7 @@ class OmnistackClusters(ResourceBase):
 class OmnistackCluster(object):
     """Implements features available for single OmniStack cluster resource."""
 
-    OBJECT_TYPE = 'omnistack_cluster'
+    OBJECT_TYPE = "omnistack_cluster"
 
     def __init__(self, connection, resource_client, data):
         self.data = data
@@ -132,3 +132,26 @@ class OmnistackCluster(object):
         clusters = [self._clusters.get_by_id(cluster["id"]) for cluster in connected_clusters]
 
         return clusters
+
+    def __refresh(self):
+        """Updates the omnistack cluster data."""
+        resource_uri = "{}/{}".format(URL, self.data["id"])
+        self.data = self._client.do_get(resource_uri)[self.OBJECT_TYPE]
+
+    def set_time_zone(self, time_zone, timeout=-1):
+        """ Sets the time zone for a cluster.
+
+        Args:
+            time_zone: The time zone in case-sensitive region/locale format
+                       for example, "America/New_York"
+            timeout : Time out for the request in seconds.
+
+        Returns:
+            object: omnistack cluster object.
+        """
+
+        method_url = "{}/{}/set_time_zone".format(URL, self.data["id"])
+        data = {"time_zone": time_zone}
+        self._client.do_post(method_url, data, timeout)
+        self.__refresh()
+        return self